+++ src/shared/decorators/index.ts
@@ -1,3 +1,3 @@
 export * from './ApiErrorResponse';
 export * from './ApiOkPaginatedResponse';
-export * from './ApiOkResponse';
+export * from './ApiSuccessResponse';
